Overhaul raw context lexing
It actually needs to use the normal expansion mechanism, not raw mode; this ends up simplifying the
logic, as there is no more need to manually perform expansions when they contain a delimiter.

The main issue was that raw args could begi inside of an expansion, but continue outside of it,
which is not what the raw-text mechanism is for.